contact-form/requirements.txt
uvicorn
fastapi
sendgrid
python-dotenv
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi